+++ lib/src/purchase/attach_payment_method_options.dart
@@ -1,0 +1,13 @@
+import 'package:revenipe_flutter/src/purchase/purchase_method.dart';
+
+class AttachPaymentMethodOptions {
+  final RevenipePurchaseMethod method;
+  final String? successUrl;
+  final String? cancelUrl;
+
+  const AttachPaymentMethodOptions({
+    required this.method,
+    this.successUrl,
+    this.cancelUrl,
+  });
+}