+++ lib/src/purchase/purchase_method.dart
@@ -2,7 +2,7 @@ enum RevenipePurchaseMethod {
   paymentSheet('payment_sheet'),
   hostedCheckout('hosted_checkout');
 
- final String value;
+  final String value;
 
- const RevenipePurchaseMethod(this.value);
+  const RevenipePurchaseMethod(this.value);
 }