+++ lib/src/services/app_service.dart
@@ -1,6 +1,5 @@
 import 'package:revenipe_flutter/revenipe_flutter.dart';
 import 'package:revenipe_flutter/src/core/respponses/app_products_response.dart';
-import 'package:revenipe_flutter/src/core/respponses/track_respopnse.dart';
 import '../network/revenipe_http_client.dart';
 
 class AppService {
@@ -13,14 +12,16 @@ class AppService {
   Future<AppProductsResponse> getProducts() {
     return _client.get<AppProductsResponse>(
       path: '${_clientBasePath}products',
-      parser: (data) => AppProductsResponse.fromJson(data as Map<String, dynamic>),
+      parser: (data) =>
+          AppProductsResponse.fromJson(data as Map<String, dynamic>),
     );
   }
 
   Future<AppProductsResponse> startPurchase() {
     return _client.get<AppProductsResponse>(
       path: '${_clientBasePath}products',
-      parser: (data) => AppProductsResponse.fromJson(data as Map<String, dynamic>),
+      parser: (data) =>
+          AppProductsResponse.fromJson(data as Map<String, dynamic>),
     );
   }
 }