+++ lib/src/services/customer_service.dart
@@ -1,5 +1,5 @@
 import 'package:revenipe_flutter/revenipe_flutter.dart';
-import 'package:revenipe_flutter/src/core/respponses/track_respopnse.dart';
+import 'package:revenipe_flutter/src/core/respponses/refresh_entitlements_response.dart';
 import '../network/revenipe_http_client.dart';
 
 class CustomerService {
@@ -7,16 +7,20 @@ class CustomerService {
 
   final RevenipeHttpClient _client;
 
-  static const String _clientBasePath = 'v1/customer/usage/';
+  static const String _clientBasePath = 'v1/customer/';
 
-  Future<TrackEntitlementResponse> track(RevenipeSession session, String entitlementId, int value) {
-
+  Future<TrackEntitlementResponse> track(
+    RevenipeSession session,
+    String entitlementId,
+    int value,
+  ) {
     var customer = session.customer;
 
-    var result = customer.entitlements.where((item) => item.entitlementId == entitlementId);
+    var result = customer.entitlements.where(
+      (item) => item.entitlementId == entitlementId,
+    );
 
     var entitlement = result.isNotEmpty ? result.first : null;
-
 
     return _client.post<TrackEntitlementResponse>(
       path: '${_clientBasePath}track',
@@ -24,9 +28,22 @@ class CustomerService {
         'access_source_id': entitlement?.accessSourceId,
         'entitlement_id': entitlementId,
         'client_id': session.customerId,
-        'value': value
+        'value': value,
       },
-      parser: (data) => TrackEntitlementResponse.fromJson(data as Map<String, dynamic>),
+      parser: (data) =>
+          TrackEntitlementResponse.fromJson(data as Map<String, dynamic>),
+    );
+  }
+
+  Future<RefreshClientEntitlementsResponse> refreshEntitlements(
+    String customerId,
+  ) {
+    return _client.post<RefreshClientEntitlementsResponse>(
+      path: '${_clientBasePath}refresh_entitlements',
+      data: <String, dynamic>{'client_id': customerId},
+      parser: (data) => RefreshClientEntitlementsResponse.fromJson(
+        data as Map<String, dynamic>,
+      ),
     );
   }
 }